requirements.txt
fastapi
uvicorn[standard]
requests
SQLAlchemy
pymysql
python-dotenv
Pillow
pydantic